fixing typos

+++ src/Personal.cpp
@@ -55,8 +55,7 @@ double Driver::calcBase(double payment, int time) {
 double Driver::calcBonus() {
     if (getWorktime() > 40) {
         return (getWorktime() * (getPayment() * 1.5));
-    }
-    else {
+    } else {
         return 0.0;
     }
 }
@@ -92,9 +91,8 @@ double Cleaner::calcBase(double payment, int time) {
 
 double Cleaner::calcBonus() {
     if (getWorktime() > 30) {
-        return (double(getWorktime()) * (getPayment() * 1.2));
-    }
-    else {
+        return (getWorktime() * (getPayment() * 1.2));
+    } else {
         return 0.0;
     }
 }
